// src/app/api/model/breakType.ts
/**
 * Goaldone API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Typ der Pause: - `ONE_TIME`: einmalige Pause an einem konkreten Datum (`date`-Feld). - `RECURRING`: unbegrenzt wiederkehrende Pause nach Recurrence-Regel –   gilt für immer, solange sie nicht gelöscht wird. - `BOUNDED_RECURRING`: wiederkehrende Pause, die nur im Zeitraum   [`validFrom`, `validUntil`] gilt. Außerhalb dieses Zeitraums   wird die Pause vom Algorithmus ignoriert. 
 */
export const BreakType = {
    OneTime: 'ONE_TIME',
    Recurring: 'RECURRING',
    BoundedRecurring: 'BOUNDED_RECURRING'
} as const;
export type BreakType = typeof BreakType[keyof typeof BreakType];